+++ src/commands/metadata/namespace.ts
@@ -16,7 +16,7 @@
  *
  */
 import { command, metadata, option } from 'clime';
-import { Metadata } from 'symbol-sdk';
+import { Metadata, Page } from 'symbol-sdk';
 
 import { ProfileCommand } from '../../interfaces/profile.command';
 import { ProfileOptions } from '../../interfaces/profile.options';
@@ -47,18 +47,18 @@ export default class extends ProfileCommand {
 
         this.spinner.start();
         const metadataHttp = profile.repositoryFactory.createMetadataRepository();
-        metadataHttp.getNamespaceMetadata(namespaceId).subscribe(
-            (metadataEntries) => {
+        metadataHttp.search({ targetId: namespaceId }).subscribe(
+            (metadataEntries: Page<Metadata>) => {
                 this.spinner.stop();
-                if (metadataEntries.length > 0) {
-                    metadataEntries.map((entry: Metadata) => {
+                if (metadataEntries.pageSize > 0) {
+                    metadataEntries.data.map((entry: Metadata) => {
                         console.log(new MetadataEntryTable(entry.metadataEntry).toString());
                     });
                 } else {
                     console.log(FormatterService.error('The namespace does not have metadata entries assigned'));
                 }
             },
-            (err) => {
+            (err: any) => {
                 this.spinner.stop();
                 console.log(FormatterService.error(err));
             },