+++ src/commands/restriction/mosaicaddress.ts
@@ -18,6 +18,7 @@
 import * as Table from 'cli-table3';
 import { HorizontalTable } from 'cli-table3';
 import { command, metadata, option } from 'clime';
+import { MosaicAddressRestriction, MosaicGlobalRestriction, MosaicGlobalRestrictionItem, Page } from 'symbol-sdk';
 
 import { ProfileCommand } from '../../interfaces/profile.command';
 import { ProfileOptions } from '../../interfaces/profile.options';
@@ -42,14 +43,16 @@ export class CommandOptions extends ProfileOptions {
 export class MosaicAddressRestrictionsTable {
     private readonly table: HorizontalTable;
 
-    constructor(public readonly mosaicAddressRestrictions: Map<string, string>) {
+    constructor(public readonly mosaicAddressRestrictions: Page<MosaicAddressRestriction | MosaicGlobalRestriction>) {
         this.table = new Table({
             style: { head: ['cyan'] },
             head: ['Type', 'Value'],
         }) as HorizontalTable;
 
-        mosaicAddressRestrictions.forEach((value: string, key: string) => {
-            this.table.push(['Key', key], ['Value', value]);
+        mosaicAddressRestrictions.data.forEach((mosaicRestriction) => {
+            mosaicRestriction.restrictions.forEach((value: string | MosaicGlobalRestrictionItem, key: string) => {
+                this.table.push(['Key', key], ['Value', value.toString()]);
+            });
         });
     }
 
@@ -77,16 +80,16 @@ export default class extends ProfileCommand {
 
         this.spinner.start();
         const restrictionHttp = profile.repositoryFactory.createRestrictionMosaicRepository();
-        restrictionHttp.getMosaicAddressRestriction(mosaicId, address).subscribe(
-            (mosaicRestrictions) => {
+        restrictionHttp.searchMosaicRestrictions({ mosaicId, targetAddress: address }).subscribe(
+            (mosaicRestrictions: Page<MosaicAddressRestriction | MosaicGlobalRestriction>) => {
                 this.spinner.stop();
-                if (mosaicRestrictions.restrictions.size > 0) {
-                    console.log(new MosaicAddressRestrictionsTable(mosaicRestrictions.restrictions).toString());
+                if (mosaicRestrictions.pageSize > 0) {
+                    console.log(new MosaicAddressRestrictionsTable(mosaicRestrictions).toString());
                 } else {
                     console.log(FormatterService.error('The address does not have mosaic address restrictions assigned'));
                 }
             },
-            (err) => {
+            (err: any) => {
                 this.spinner.stop();
                 console.log(FormatterService.error(err));
             },