+++ src/commands/transaction/mosaicmetadata.ts
@@ -16,7 +16,7 @@
  *
  */
 import { command, metadata, option } from 'clime';
-import { Deadline, MetadataTransactionService, MetadataType } from 'symbol-sdk';
+import { Deadline, MetadataTransactionService } from 'symbol-sdk';
 
 import { AnnounceTransactionsCommand } from '../../interfaces/announce.transactions.command';
 import { AnnounceTransactionsOptions } from '../../interfaces/announce.transactions.options';
@@ -82,15 +82,14 @@ export default class extends AnnounceTransactionsCommand {
         const metadataHttp = profile.repositoryFactory.createMetadataRepository();
         const metadataTransactionService = new MetadataTransactionService(metadataHttp);
         const metadataTransaction = await metadataTransactionService
-            .createMetadataTransaction(
+            .createMosaicMetadataTransaction(
                 Deadline.create(),
                 account.networkType,
-                MetadataType.Mosaic,
                 targetAddress,
+                mosaic,
                 key,
                 value,
                 account.address,
-                mosaic,
                 maxFee,
             )
             .toPromise();